fix: correct import paths in lib/ files after move
- model-enhancer.ts: ./provider-failover/ -> ../provider-failover/
- util.ts: ./lib/logger.ts -> ./logger.ts
